+++ src/helpers/write-to-package.json.ts
@@ -1,0 +1,30 @@
+import fs from 'fs'
+import path from 'path'
+
+
+/**
+ * Adds a script or value to the package.json file
+ * If the key doesn't exist, it will be created.
+ * 
+ * @param key {string} - the key in the package.json file
+ * @param value the value to set the key to
+ */
+export function editToPackageJson(key: string, item: string | object) {
+    try {
+        const file = fs.readFileSync(path.join(process.cwd(), 'package.json'), 'utf8')
+        const line = file.split(/\r\n|\n/)[1]
+
+        // Determine tabWidth
+        let tabWidth = 4
+        if(!/    /gm.test(line) && /  /gm.test(line)) tabWidth = 2
+
+        const packageJson = JSON.parse(fs.readFileSync(path.join(process.cwd(), 'package.json'), 'utf8'))
+        if(!packageJson[key]) packageJson[key] = {}
+        Object.assign(packageJson[key], item)
+        fs.writeFileSync(path.join(process.cwd(), 'package.json'), JSON.stringify(packageJson, null, tabWidth), 'utf8')
+    } catch (error: any) {
+        console.log('error', error.message)
+    }
+    
+}
+